fix dtypes

+++ tests/python_package_test/test_cpp_python_consistency.py
@@ -26,9 +26,9 @@ class FileLoader(object):
     def load_dataset(self, suffix, is_sparse=False):
         filename = os.path.join(self.directory, self.prefix + suffix)
         if is_sparse:
-            return load_svmlight_file(filename)
+            return load_svmlight_file(filename, dtype=np.float64)
         else:
-            mat = np.loadtxt(filename)
+            mat = np.loadtxt(filename, dtype=np.float64)
             return mat[:, 1:], mat[:, 0]
 
     def load_field(self, suffix):
